Change ioctls to run through SyscallConn
As described in https://github.com/golang/go/issues/24331, calling
os.File.Fd() places the file in blocking mode, which prevents
using Close() to interrupt a Read(). This in turn means that there
is no way to cleanly shut down a goroutine that is blocked on Read().

The solution to the issue is to use the syscall.RawConn interface
via os.File.SyscallConn(), which has a Control() method that takes
a callback function that can access the file descriptor while
keeping the file in nonblocking mode.

This has been accomplished by adding the (*InputDevice).Ioctl()
receiver to look up the file descriptor from the RawConn interface
before an ioctl is executed.

With this change, blocking calls to Read() and ReadOne() can be
interrupted by calling the new Close() receiver that wraps
os.File.Close().